@@ -65,31 +66,31 @@
     "winston-transport": "^4.6.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^18.2.0",
-    "@commitlint/config-conventional": "^18.1.0",
-    "@faker-js/faker": "^8.2.0",
+    "@commitlint/cli": "^18.4.3",
+    "@commitlint/config-conventional": "^18.4.3",
+    "@faker-js/faker": "^8.3.1",
     "@jest/globals": "^29.7.0",
-    "@mikro-orm/cli": "^5.9.2",
-    "@mikro-orm/seeder": "^5.9.2",
+    "@jest/types": "^29.6.3",
     "@shoginn/discordjs-mock": "^1.0.1",
-    "@trivago/prettier-plugin-sort-imports": "^4.2.1",
+    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@tsconfig/strictest": "^2.0.2",
-    "@types/convict": "^6.1.5",
-    "@types/convict-format-with-validator": "^6.0.4",
-    "@types/jest": "^29.5.7",
-    "@types/lodash": "^4.14.200",
-    "@types/mitm": "^1.3.6",
-    "@types/node": "~20.8.10",
-    "@types/node-schedule": "^2.1.2",
-    "@typescript-eslint/eslint-plugin": "^6.10.0",
-    "@typescript-eslint/parser": "^6.10.0",
+    "@types/convict": "^6.1.6",
+    "@types/convict-format-with-validator": "^6.0.5",
+    "@types/jest": "^29.5.10",
+    "@types/lodash": "^4.14.202",
+    "@types/mitm": "^1.3.8",
+    "@types/node": "~20.10.2",
+    "@types/node-schedule": "^2.1.4",
+    "@typescript-eslint/eslint-plugin": "^6.13.1",
+    "@typescript-eslint/parser": "^6.13.1",
     "depcheck": "^1.4.7",
-    "eslint": "^8.53.0",
-    "eslint-config-prettier": "~9.0.0",
+    "eslint": "^8.55.0",
+    "eslint-config-prettier": "~9.1.0",
     "eslint-plugin-deprecation": "^2.0.0",
     "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-jest": "^27.6.0",
-    "eslint-plugin-jsdoc": "^46.8.2",
+    "eslint-plugin-jsdoc": "^46.9.0",
+    "eslint-plugin-prettier": "^5.0.1",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-unicorn": "^49.0.0",
     "genversion": "^3.1.1",
@@ -99,13 +100,14 @@
     "jest-fetch-mock": "^3.0.3",
     "mitm": "^1.7.2",
     "nyc": "^15.1.0",
-    "prettier": "^3.0.3",
+    "openapi-typescript": "~6.7.2",
+    "prettier": "^3.1.0",
     "rimraf": "^5.0.5",
     "ts-jest": "^29.1.1",
     "ts-jest-resolver": "^2.0.1",
     "ts-mockito": "^2.6.1",
     "ts-node": "^10.9.1",
-    "typescript": "5.2.2",
+    "typescript": "5.3.2",
     "typesync": "^0.11.1"
   },
   "typesync": {
