@@ -1,58 +1,83 @@
-import pkg, {
-  Account,
-  assignGroupID,
-  AtomicTransactionComposer,
-  isValidAddress,
-  makeAssetTransferTxnWithSuggestedParamsFromObject,
-  mnemonicToSecretKey,
-  Transaction,
-  TransactionType,
-  waitForConfirmation,
-} from 'algosdk';
-import SearchForTransactions from 'algosdk/dist/types/client/v2/indexer/searchForTransactions.js';
-import { SuggestedParamsWithMinFee } from 'algosdk/dist/types/types/transactions/base.js';
-import chunk from 'lodash/chunk.js';
-import isString from 'lodash/isString.js';
-import { container, injectable, singleton } from 'tsyringe';
+import * as algokit from '@algorandfoundation/algokit-utils';
+import { SigningAccount } from '@algorandfoundation/algokit-utils/types/account';
+import { AlgoConfig } from '@algorandfoundation/algokit-utils/types/network-client';
+import { TransferAssetParams } from '@algorandfoundation/algokit-utils/types/transfer';
+import { Account, TransactionType } from 'algosdk';
+import AlgodClient from 'algosdk/dist/types/client/v2/algod/algod.js';
+import IndexerClient from 'algosdk/dist/types/client/v2/indexer/indexer.js';
+import { inject, injectable, singleton } from 'tsyringe';
 
 import { getConfig } from '../config/config.js';
-import { AlgoClientEngine } from '../model/framework/engine/impl/algo-client-engine.js';
+import { PostConstruct } from '../decorators/post-construct.js';
 import {
-  AlgorandTransaction,
+  Arc69MetaData,
   Arc69Payload,
-  AssetGroupTransferOptions,
+  Asset,
   AssetHolding,
-  AssetLookupResult,
   AssetTransferOptions,
   AssetType,
-  ClaimTokenResponse,
   ClaimTokenTransferOptions,
   ClawbackTokenTransferOptions,
-  LookupAssetBalances,
-  LookupAssetBalancesResult,
-  MainAssetResult,
-  PendingTransactionResponse,
-  SearchCriteria,
+  LookupAssetBalancesResponse,
+  LookUpAssetByIDResponse,
+  MiniAssetHolding,
   TipTokenTransferOptions,
-  TransactionSearchResults,
-  UnclaimedAsset,
-  WalletWithUnclaimedAssets,
-} from '../model/types/algorand.js';
-import { AlgorandRepository } from '../repositories/algorand-repository.js';
+  TransactionResultOrError,
+} from '../types/algorand.js';
+import { WalletAddress } from '../types/core.js';
 import logger from '../utils/functions/logger-factory.js';
 
+import { handleTransferErrors } from './algorand.errorprocessor.js';
 import { CustomCache } from './custom-cache.js';
 
-const { generateAccount } = pkg;
 @singleton()
 @injectable()
-export class Algorand extends AlgoClientEngine {
-  public AlgorandRepository: AlgorandRepository;
-  public constructor(algoRepository?: AlgorandRepository) {
-    super();
-    this.AlgorandRepository = algoRepository ?? new AlgorandRepository();
+export class Algorand {
+  private _algodClient?: AlgodClient;
+  private _indexerClient?: IndexerClient;
+  private _clawbackAccount?: Account | SigningAccount;
+  private _claimTokenAccount?: Account | SigningAccount;
+  public constructor(@inject(CustomCache) private customCache: CustomCache) {}
+  setupClients(config?: AlgoConfig): void {
+    config = config || algokit.getConfigFromEnvOrDefaults();
+    if (getConfig().get().nodeEnv === 'production') {
+      config = {
+        ...config,
+        algodConfig: algokit.getAlgoNodeConfig('mainnet', 'algod'),
+        indexerConfig: algokit.getAlgoNodeConfig('mainnet', 'indexer'),
+      };
+    }
+    //TODO: remove this once things are working
+    process.env['ALGOD_SERVER'] = config.algodConfig.server;
+    process.env['ALGOD_PORT'] = config.algodConfig.port?.toString();
+    process.env['ALGOD_TOKEN'] = config.algodConfig.token?.toString();
+    this._algodClient = algokit.getAlgoClient(config.algodConfig);
+    this._indexerClient = algokit.getAlgoIndexerClient(config.indexerConfig);
   }
-
+  get algodClient(): AlgodClient {
+    if (!this._algodClient) {
+      this.setupClients();
+    }
+    return this._algodClient as AlgodClient;
+  }
+  get indexerClient(): IndexerClient {
+    if (!this._indexerClient) {
+      this.setupClients();
+    }
+    return this._indexerClient as IndexerClient;
+  }
+  get clawbackAccount(): Account | SigningAccount {
+    if (!this._clawbackAccount) {
+      throw new Error('Clawback Account not set');
+    }
+    return this._clawbackAccount;
+  }
+  get claimTokenAccount(): Account | SigningAccount {
+    if (!this._claimTokenAccount) {
+      throw new Error('Claim Token Account not set');
+    }
+    return this._claimTokenAccount;
+  }
   /**
    * Takes a note and returns the arc69 payload if it exists
    *
@@ -75,87 +100,53 @@ export class Algorand extends AlgoClientEngine {
 
     return JSON.parse(json) as Arc69Payload;
   }
-
-  /**
-   *Validates wallet address
-   *
-   * @param {string} walletAddress
-   * @returns {*} boolean
-   * @memberof Algorand
-   */
-  validateWalletAddress(walletAddress: string): boolean {
-    return isValidAddress(walletAddress);
-  }
-
-  /**
-   * Creates a new wallet using the Algorand SDK
-   *
-   * @returns {*}  {string}
-   * @memberof Algorand
-   */
-  generateWalletAccount(): string {
-    const account = generateAccount();
-    return account.addr;
-  }
-  /**
-     * Get account from mnemonic
-     *
-
-     * @param {string} mnemonic
-     * @returns {*}  {Account}
-     */
-  getAccountFromMnemonic(mnemonic: string): Account | undefined {
-    if (!isString(mnemonic)) {
+  @PostConstruct
+  async initAccounts(): Promise<void> {
+    if (this._claimTokenAccount && this._clawbackAccount) {
       return;
     }
-    const cleanedMnemonic = mnemonic
-      .replaceAll(/\W/g, ' ')
-      .replaceAll(/\s{2,}/g, ' ')
-      .trimEnd()
-      .trimStart();
-    let secretKey: Account;
-    try {
-      secretKey = mnemonicToSecretKey(cleanedMnemonic);
-    } catch (error) {
-      logger.error(`Failed to get account from mnemonic: ${JSON.stringify(error)}`);
-      return;
-    }
-    return secretKey;
+    const { token, clawback } = await this.getMnemonicAccounts();
+    this._claimTokenAccount = token;
+    this._clawbackAccount = clawback;
   }
-
   /**
    * Retrieves the mnemonic from the environment
    * If the claim mnemonic is not set then it will use the clawback mnemonic
    *
-   * @param {string} [clawBackTokenMnemonic]
-   * @param {string} [claimTokenMnemonic]
    * @returns {*}  {{ token: Account; clawback: Account }}
    * @memberof Algorand
    */
-  getMnemonicAccounts(
-    clawBackTokenMnemonic?: string,
-    claimTokenMnemonic?: string,
-  ): { token: Account; clawback: Account } {
-    const { clawbackTokenMnemonic: configClawback, claimTokenMnemonic: configClaim } =
-      getConfig().get();
-    clawBackTokenMnemonic = clawBackTokenMnemonic || configClawback;
-    claimTokenMnemonic = claimTokenMnemonic || configClaim;
-    const claimTokenAccount = claimTokenMnemonic
-      ? this.getAccountFromMnemonic(claimTokenMnemonic)
-      : this.getAccountFromMnemonic(clawBackTokenMnemonic);
-
-    const clawbackAccount = this.getAccountFromMnemonic(clawBackTokenMnemonic);
-
-    if (!claimTokenAccount || !clawbackAccount) {
-      throw new Error('Failed to get accounts from mnemonics');
+  async getMnemonicAccounts(): Promise<{
+    token: Account | SigningAccount;
+    clawback: Account | SigningAccount;
+  }> {
+    let claimTokenAccount;
+    let clawbackTokenAccount;
+    try {
+      clawbackTokenAccount = await algokit.mnemonicAccountFromEnvironment(
+        'clawback_token',
+        this.algodClient,
+      );
+    } catch (error) {
+      throw new Error(`Clawback Token not set: ${(error as Error).message}`);
     }
 
-    return { token: claimTokenAccount, clawback: clawbackAccount };
+    try {
+      claimTokenAccount = await algokit.mnemonicAccountFromEnvironment(
+        'claim_token',
+        this.algodClient,
+      );
+    } catch {
+      claimTokenAccount = clawbackTokenAccount;
+      logger.warn('Claim Token not set, using Clawback Token');
+    }
+
+    return { token: claimTokenAccount, clawback: clawbackTokenAccount };
   }
-  private createCacheKey(walletAddress: string, assetType: AssetType): string {
+  private createCacheKey(walletAddress: WalletAddress, assetType: AssetType): string {
     return `walletAccountAssets-${walletAddress}-${assetType}`;
   }
-  private logAssets<T>(assetType: AssetType, walletAddress: string, assets: T[]): void {
+  private logAssets<T>(assetType: AssetType, walletAddress: WalletAddress, assets: T[]): void {
     if (assets.length === 0) {
       logger.info(`Didn't find any ${assetType} for account ${walletAddress}`);
     } else if (assetType === AssetType.CreatedAssets) {
@@ -164,13 +155,12 @@ export class Algorand extends AlgoClientEngine {
   }
 
   private async getCachedOrFetch<T>(cacheKey: string, fetchFunction: () => Promise<T>): Promise<T> {
-    const cache = container.resolve(CustomCache);
-    const cached = (await cache.get(cacheKey)) as T;
+    const cached = (await this.customCache.get(cacheKey)) as T;
     if (cached) {
       return cached;
     }
     const data = await fetchFunction();
-    cache.set(cacheKey, data); // setting cache for 1 hour
+    this.customCache.set(cacheKey, data); // setting cache for 1 hour
     return data;
   }
   /**
@@ -182,8 +172,8 @@ export class Algorand extends AlgoClientEngine {
    * @returns {*}  {(Promise<AssetHolding[] | MainAssetResult[] | []>)}
    * @memberof Algorand
    */
-  public async getAccountAssets<T extends AssetHolding | MainAssetResult>(
-    walletAddress: string,
+  public async getAccountAssets<T extends AssetHolding | Asset>(
+    walletAddress: WalletAddress,
     assetType: AssetType,
   ): Promise<T[] | []> {
     const cacheKey = this.createCacheKey(walletAddress, assetType);
@@ -195,13 +185,43 @@ export class Algorand extends AlgoClientEngine {
   }
 
   /**
+   * Retrieves the asset holdings for a given wallet address and asset index.
+   * @param {string} walletAddress - The wallet address to retrieve asset holdings for.
+   * @param {number} assetIndex - The index of the asset to retrieve holdings for.
+   * @returns {*} {Promise<AssetHolding[] | []>}
+   * @memberof Algorand
+   */
+  public async getHeldAssetFromAccount(
+    walletAddress: WalletAddress,
+    assetIndex: number,
+  ): Promise<AssetHolding | undefined> {
+    const request = this.algodClient.accountAssetInformation(walletAddress, assetIndex);
+    let accountAssets;
+    try {
+      accountAssets = await request.do();
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } catch (error: any) {
+      if (('response' in error && Number(error.response.status) === 404) || error.status === 404) {
+        return;
+      }
+      logger.error(
+        `Failed to get asset holdings for asset ${assetIndex} and wallet ${walletAddress}`,
+      );
+      logger.error(error);
+      return;
+    }
+    const assets = accountAssets['asset-holding'];
+    return assets;
+  }
+
+  /**
    * Retrieves the assets owned by a wallet address
    *
    * @param {string} walletAddress
    * @returns {*}  {Promise<AssetHolding[]>}
    * @memberof Algorand
    */
-  public async lookupAssetsOwnedByAccount(walletAddress: string): Promise<AssetHolding[]> {
+  public async lookupAssetsOwnedByAccount(walletAddress: WalletAddress): Promise<AssetHolding[]> {
     return await this.getAccountAssets<AssetHolding>(walletAddress, AssetType.Assets);
   }
 
@@ -209,53 +229,45 @@ export class Algorand extends AlgoClientEngine {
    * Retrieves the assets created by a wallet address
    *
    * @param {string} walletAddress
-   * @returns {*}  {Promise<MainAssetResult[]>}
+   * @returns {*}  {Promise<Asset[]>}
    * @memberof Algorand
    */
-  public async getCreatedAssets(walletAddress: string): Promise<MainAssetResult[]> {
-    return await this.getAccountAssets<MainAssetResult>(walletAddress, AssetType.CreatedAssets);
+  public async getCreatedAssets(walletAddress: WalletAddress): Promise<Asset[]> {
+    return await this.getAccountAssets<Asset>(walletAddress, AssetType.CreatedAssets);
   }
   /**
    * Get the token opt in status for a wallet address
    *
    * @param {string} walletAddress
-   * @param {number} optInAssetId
+   * @param {number} assetIndex
    * @returns {*}  {(Promise<{ optedIn: boolean; tokens: number | bigint }>)}
    * @memberof Algorand
    */
   async getTokenOptInStatus(
-    walletAddress: string,
-    optInAssetId: number,
+    walletAddress: WalletAddress,
+    assetIndex: number,
   ): Promise<{ optedIn: boolean; tokens: number | bigint }> {
-    const accountAssets = await this.getAccountAssets<AssetHolding>(
-      walletAddress,
-      AssetType.Assets,
-    );
-    const asset = accountAssets.find((a) => a['asset-id'] === optInAssetId);
-
-    return {
-      optedIn: !!asset,
-      tokens: asset?.amount || 0,
-    };
+    const accountAssets = await this.getHeldAssetFromAccount(walletAddress, assetIndex);
+    const tokens = accountAssets?.amount ?? 0;
+    const optedIn = !!accountAssets;
+    return { optedIn, tokens };
   }
   async lookupAssetBalances(
-    assetId: number,
+    assetIndex: number,
     getAll?: boolean | undefined,
-  ): Promise<Array<Pick<LookupAssetBalances, 'address' | 'amount'>>> {
-    let holders: Array<Pick<LookupAssetBalances, 'address' | 'amount'>> = [];
+  ): Promise<MiniAssetHolding[]> {
+    let holders: MiniAssetHolding[] = [];
     let nextToken: string | undefined;
     do {
-      const response = this.indexerClient.lookupAssetBalances(assetId).includeAll(getAll);
+      const response = this.indexerClient.lookupAssetBalances(assetIndex).includeAll(getAll);
       if (nextToken) {
         response.nextToken(nextToken);
       }
-      const result = (await response.do()) as LookupAssetBalancesResult;
+      const result = (await response.do()) as LookupAssetBalancesResponse;
 
       holders = [
         ...holders,
-        ...(result['balances']
-          ?.filter((balance) => balance.amount > 0)
-          .map((balance) => ({ address: balance.address, amount: balance.amount })) || []),
+        ...(result['balances']?.filter((balance) => balance.amount > 0) || []),
       ];
       nextToken = result?.['next-token'];
     } while (nextToken !== undefined);
@@ -264,31 +276,17 @@ export class Algorand extends AlgoClientEngine {
   /**
    * Get the asset by index
    *
-   * @param {number} index
+   * @param {number} assetIndex
    * @param {(boolean | undefined)} [getAll]
    * @returns {*}  {Promise<AssetLookupResult>}
    * @memberof Algorand
    */
   async lookupAssetByIndex(
-    index: number,
+    assetIndex: number,
     getAll?: boolean | undefined,
-  ): Promise<AssetLookupResult> {
-    const request = this.indexerClient.lookupAssetByID(index).includeAll(getAll);
-    return (await request.do()) as AssetLookupResult;
-  }
-
-  /**
-   * Asynchronously searches for transactions based on the provided search criteria.
-   *
-   * @template T - The type of the result returned by the search.
-   *
-   * @param {SearchCriteria<SearchForTransactions>} searchCriteria - The search criteria to be used.
-   *
-   * @returns {Promise<T>} - A promise that resolves to the search result of type T.
-   */
-  async searchTransactions<T>(searchCriteria: SearchCriteria<SearchForTransactions>): Promise<T> {
-    const request = searchCriteria(this.indexerClient.searchForTransactions());
-    return (await request.do()) as T;
+  ): Promise<LookUpAssetByIDResponse> {
+    const request = this.indexerClient.lookupAssetByID(assetIndex).includeAll(getAll);
+    return (await request.do()) as LookUpAssetByIDResponse;
   }
 
   /**
@@ -300,7 +298,7 @@ export class Algorand extends AlgoClientEngine {
    */
   async getAssetArc69Metadata(assetIndex: number): Promise<Arc69Payload | undefined> {
     try {
-      const acfgTransactions = await this.searchTransactions<TransactionSearchResults>((s) =>
+      const acfgTransactions = await algokit.searchTransactions(this.indexerClient, (s) =>
         s.assetID(assetIndex).txType(TransactionType.acfg),
       );
       // Sort the transactions by round number in descending order
@@ -322,11 +320,9 @@ export class Algorand extends AlgoClientEngine {
    * Retrieves the bulk asset metadata for the given asset indexes.
    *
    * @param {number[]} assetIndexes - The indexes of the assets to retrieve metadata for.
-   * @returns {Promise<{ id: number; arc69: Arc69Payload }[]>} - The bulk asset metadata containing the asset ID and ARC69 payload.
+   * @returns {Promise<Arc69MetaData[]>} - The bulk asset metadata containing the asset ID and ARC69 payload.
    */
-  async getBulkAssetArc69Metadata(
-    assetIndexes: number[],
-  ): Promise<Array<{ id: number; arc69: Arc69Payload }>> {
+  async getBulkAssetArc69Metadata(assetIndexes: number[]): Promise<Arc69MetaData[] | []> {
     const assetMeta = await Promise.all(
       assetIndexes.map(async (assetId) => {
         const arc69Metadata = await this.getAssetArc69Metadata(assetId);
@@ -342,191 +338,51 @@ export class Algorand extends AlgoClientEngine {
     }>;
   }
 
-  async unclaimedGroupClaim(
-    chunk: WalletWithUnclaimedAssets[],
-    asset: UnclaimedAsset,
-  ): Promise<void> {
-    try {
-      const claimStatus = await this.groupClaimToken({
-        assetIndex: asset.id,
-        groupTransfer: chunk,
-      });
-      const chunkUnclaimedAssets = chunk.reduce(
-        (accumulator, current) => accumulator + current.unclaimedTokens,
-        0,
-      );
-
-      if (claimStatus.txId) {
-        logger.info(
-          `Auto Claimed ${
-            chunk.length
-          } wallets with a total of ${chunkUnclaimedAssets.toLocaleString()} ${
-            asset.name
-          } -- Block: ${claimStatus?.status?.['confirmed-round'] ?? 'unk'} -- TxId: ${
-            claimStatus.txId
-          }`,
-        );
-        // Remove the unclaimed tokens from the wallet
-        await this.AlgorandRepository.removeUnclaimedTokensFromMultipleWallets(chunk, asset);
-      } else {
-        throw new Error(
-          `Auto Claim Failed ${
-            chunk.length
-          } wallets with a total of ${chunkUnclaimedAssets.toLocaleString()} ${asset.name}`,
-        );
-      }
-    } catch (error) {
-      logger.error(`Auto Claim Failed: ${(error as Error).message}`);
-
-      this.logFailedChunkedWallets(chunk);
-    }
+  claimToken(options: ClaimTokenTransferOptions): Promise<TransactionResultOrError> {
+    return this.transferOptionsProcessor(options);
   }
-  async batchTransActionProcessor(
-    walletsWithUnclaimedAssets: WalletWithUnclaimedAssets[],
-    asset: UnclaimedAsset,
-  ): Promise<void> {
-    if (walletsWithUnclaimedAssets.length === 0) {
-      return;
-    }
-
-    // Only 16 wallets can be claimed in a single atomic transfer so we need to split the array into chunks
-    const arraySize = AtomicTransactionComposer.MAX_GROUP_SIZE;
-    const chunkedWallets = chunk(walletsWithUnclaimedAssets, arraySize);
-    const promiseArray = [];
-    logger.info(
-      `Claiming ${walletsWithUnclaimedAssets.length} wallets with unclaimed ${asset.name}...`,
-    );
-    logger.info(
-      `For a total of ${walletsWithUnclaimedAssets
-        .reduce((accumulator, current) => accumulator + current.unclaimedTokens, 0)
-        .toLocaleString()} ${asset.name}`,
-    );
-    for (const chunk of chunkedWallets) {
-      // sum the total unclaimed Assets for all users using [1] in tuple
-      // Claim all unclaimed Assets using atomic transfer
-      promiseArray.push(this.unclaimedGroupClaim(chunk, asset));
-    }
-    await Promise.all(promiseArray);
+  tipToken(options: TipTokenTransferOptions): Promise<TransactionResultOrError> {
+    return this.transferOptionsProcessor(options);
   }
 
-  private logFailedChunkedWallets(chunk: WalletWithUnclaimedAssets[]): void {
-    for (const wallet of chunk) {
-      logger.error(`${wallet.walletAddress} -- ${wallet.unclaimedTokens} -- ${wallet.userId}`);
-    }
-  }
-  async checkSenderBalance(
-    senderAddress: string,
-    optInAssetId: number,
-    amount: number,
-  ): Promise<number | bigint> {
-    const { tokens: senderBalance } = await this.getTokenOptInStatus(senderAddress, optInAssetId);
-    if (senderBalance < amount) {
-      throw new Error('Insufficient Funds to cover transaction');
-    }
-    return senderBalance;
-  }
-  async getSuggestedParameters(): Promise<SuggestedParamsWithMinFee> {
-    return await this.algodClient.getTransactionParams().do();
-  }
-
-  async claimToken(options: ClaimTokenTransferOptions): Promise<ClaimTokenResponse> {
-    return await this.transferOptionsProcessor(options);
-  }
-  async tipToken(options: TipTokenTransferOptions): Promise<ClaimTokenResponse> {
-    return await this.transferOptionsProcessor(options);
-  }
-
-  async purchaseItem(options: ClawbackTokenTransferOptions): Promise<ClaimTokenResponse> {
+  purchaseItem(options: ClawbackTokenTransferOptions): Promise<TransactionResultOrError> {
     options.clawback = true;
-    return await this.transferOptionsProcessor(options);
-  }
-  async groupClaimToken(options: AssetGroupTransferOptions): Promise<ClaimTokenResponse> {
-    return await this.transferOptionsProcessor(options);
+    return this.transferOptionsProcessor(options);
   }
 
-  async transferOptionsProcessor(options: AssetTransferOptions): Promise<ClaimTokenResponse> {
+  async transferOptionsProcessor(
+    options: AssetTransferOptions,
+    note: string = 'This transaction was made by the Daruma Game Bot On Discord!',
+  ): Promise<TransactionResultOrError> {
+    const transferAssetParameters = this.buildTransferAssetParams(options, note);
+    return await this.transferAsset(transferAssetParameters);
+  }
+  async transferAsset(
+    transferAssetParameters: TransferAssetParams,
+  ): Promise<TransactionResultOrError> {
     try {
-      if (options.senderAddress && options.amount) {
-        await this.checkSenderBalance(options.senderAddress, options.assetIndex, options.amount);
-      }
-      return await this.assetTransfer(options);
+      return await algokit.transferAsset(transferAssetParameters, this.algodClient);
     } catch (error) {
-      const errorMessage = 'Failed the Token transfer';
-      return this.claimErrorProcessor(error, errorMessage);
+      return { error: true, message: handleTransferErrors(error) };
     }
   }
-  claimErrorProcessor(error: unknown, errorMessage: string): ClaimTokenResponse {
-    if (error instanceof Error) {
-      logger.error(error.stack);
-      errorMessage += `: ${error.message}`;
+  getSignerAccount(options: AssetTransferOptions): Account | SigningAccount {
+    if (options.senderAddress) {
+      return this.clawbackAccount;
     }
-    logger.error(errorMessage);
-    return { error: errorMessage };
+    return this.claimTokenAccount;
   }
-
-  async assetTransfer(options: AssetTransferOptions): Promise<ClaimTokenResponse> {
-    const { assetIndex, amount, receiverAddress, senderAddress, groupTransfer, clawback } = options;
-    const suggestedParameters = await this.getSuggestedParameters();
-    const { token: claimTokenAccount, clawback: clawbackAccount } = this.getMnemonicAccounts();
-    const fromAccount = senderAddress ? claimTokenAccount : clawbackAccount;
-    const transferOptions: AlgorandTransaction = {
-      from: fromAccount.addr,
-      to: clawback ? clawbackAccount.addr : receiverAddress ?? '',
-      amount: amount ?? 0,
-      assetIndex,
-      suggestedParams: suggestedParameters,
+  buildTransferAssetParams(options: AssetTransferOptions, note?: string): TransferAssetParams {
+    const { assetIndex, amount, receiverAddress, senderAddress, clawback } = options;
+    const fromAccount = this.getSignerAccount(options);
+    const transferOptions: TransferAssetParams = {
+      from: fromAccount,
+      to: clawback ? this.clawbackAccount : receiverAddress!,
+      amount: amount,
+      assetId: assetIndex,
+      clawbackFrom: senderAddress!,
+      note,
     };
-    if (senderAddress) {
-      transferOptions.revocationTarget = senderAddress;
-    }
-    let rawSignedTxn: Uint8Array | Uint8Array[];
-
-    if (groupTransfer) {
-      const rawMultiTxn = this.makeMultipleAssetTransferTransaction(transferOptions, groupTransfer);
-      rawSignedTxn = rawMultiTxn.map((txn) => txn.signTxn(fromAccount.sk));
-    } else {
-      const singleTxn = this.makeSingleAssetTransferTransaction(transferOptions);
-      rawSignedTxn = singleTxn.signTxn(fromAccount.sk);
-    }
-    const rawTxn = await this.algodClient.sendRawTransaction(rawSignedTxn).do();
-
-    const confirmationStatus = (await waitForConfirmation(
-      this.algodClient,
-      rawTxn.txId,
-      5,
-    )) as PendingTransactionResponse;
-    return { txId: rawTxn?.txId, status: confirmationStatus };
-  }
-  makeSingleAssetTransferTransaction(options: AlgorandTransaction): Transaction {
-    return makeAssetTransferTxnWithSuggestedParamsFromObject(options);
-  }
-  makeMultipleAssetTransferTransaction(
-    options: AlgorandTransaction,
-    groupTransfer: WalletWithUnclaimedAssets[],
-  ): Transaction[] {
-    const rawMultiTxn: Transaction[] = [];
-    for (const address of groupTransfer) {
-      const txnObject: AlgorandTransaction = {
-        ...options,
-        to: address.walletAddress,
-        amount: address.unclaimedTokens,
-      };
-      rawMultiTxn.push(this.makeSingleAssetTransferTransaction(txnObject));
-    }
-    // Assign the group id to the multi signed transaction
-    return assignGroupID(rawMultiTxn);
-  }
-  async unclaimedAutomated(claimThreshold: number, asset: UnclaimedAsset): Promise<void> {
-    const walletsWithUnclaimedAssets =
-      await this.AlgorandRepository.fetchWalletsWithUnclaimedAssets(claimThreshold, asset);
-    if (walletsWithUnclaimedAssets.length === 1 && walletsWithUnclaimedAssets[0]) {
-      await this.claimToken({
-        assetIndex: asset.id,
-        amount: walletsWithUnclaimedAssets[0].unclaimedTokens,
-        receiverAddress: walletsWithUnclaimedAssets[0].walletAddress,
-      });
-      return;
-    }
-    await this.batchTransActionProcessor(walletsWithUnclaimedAssets, asset);
+    return transferOptions;
   }
 }
