@@ -1,12 +1,14 @@
 export default {
   transform: {
-    '^.+\\.tsx?$': [
+    '<regex_match_files>': [
       'ts-jest',
       {
-        tsconfig: '<rootDir>/tests/tsconfig.json',
+        tsconfig: '<rootDir>/tsconfig.test.json',
       },
     ],
   },
+  coveragePathIgnorePatterns: ['tests'],
+  testEnvironment: 'node',
   workerIdleMemoryLimit: 0.2,
   testTimeout: 30_000,
   preset: 'ts-jest/presets/default-esm',
@@ -18,16 +20,5 @@ export default {
   moduleDirectories: ['node_modules', '__mocks__'],
   transformIgnorePatterns: [],
   testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/build/'],
-  setupFiles: ['<rootDir>/tests/utils/setup.ts', '<rootDir>/tests/utils/jest-mitm.ts'],
-  collectCoverageFrom: ['<rootDir>/src/**/*.{js,jsx,ts,tsx}'],
-  coveragePathIgnorePatterns: [
-    '<rootDir>/.*\\.d\\.ts$',
-    '<rootDir>/src/enums/.*',
-    '<rootDir>/src/events/.*',
-    '<rootDir>/src/guards/.*',
-    '<rootDir>/src/main\\.ts$',
-    '<rootDir>/src/mikro-orm\\.config\\.ts$',
-    '<rootDir>/src/model/framework/decorators/.*',
-    '<rootDir>/src/model/types/.*',
-  ],
+  setupFiles: ['<rootDir>/tests/setup/setup.ts'],
 };
