@@ -6,15 +6,13 @@ import type { ArgsOf } from 'discordx';
 import { injectable } from 'tsyringe';
 
 import { Maintenance } from '../guards/maintenance.js';
+import { InteractionUtils } from '../utils/classes/interaction-utils.js';
 import logger from '../utils/functions/logger-factory.js';
-import { syncUser, syncUserMongo } from '../utils/functions/synchronizer.js';
-import { getDeveloperMentions, InteractionUtils } from '../utils/utils.js';
+import { getDeveloperMentions } from '../utils/functions/owner-utils.js';
 
 @Discord()
 @injectable()
 export default class InteractionCreateEvent {
-  constructor() {}
-
   @On()
   @Guard(Maintenance)
   async interactionCreate(
@@ -23,9 +21,6 @@ export default class InteractionCreateEvent {
   ): Promise<void> {
     try {
       // insert user in db if not exists
-      await syncUser(interaction.user);
-      await syncUserMongo(interaction.user);
-      // update last interaction time of both user and guild
       await client.executeInteraction(interaction);
     } catch (error) {
       if (error instanceof Error) {
