@@ -2,15 +2,16 @@ import { CommandInteraction } from 'discord.js';
 
 import { Client, Next } from 'discordx';
 
-import { InteractionUtils, isDeveloper } from '../utils/utils.js';
+import { InteractionUtils } from '../utils/classes/interaction-utils.js';
+import { isDeveloper } from '../utils/functions/owner-utils.js';
 
 export function BotOwnerOnly(
   argument: CommandInteraction,
   _client: Client,
   next: Next,
 ): Promise<unknown> {
-  const userId = argument?.user?.id;
-  if (!isDeveloper(userId)) {
+  const discordUserId = argument?.user?.id;
+  if (!isDeveloper(discordUserId)) {
     return InteractionUtils.replyOrFollowUp(argument, {
       ephemeral: true,
       content: 'You are not the bot owner!\nCommand aborted',
