@@ -15,8 +15,10 @@ import {
 
 import { ArgsOf, GuardFunction, SimpleCommandMessage } from 'discordx';
 
-import { isInMaintenance } from '../utils/functions/maintenance.js';
-import { isDeveloper } from '../utils/utils.js';
+import { container } from 'tsyringe';
+
+import { MaintenanceService } from '../services/maintenance.js';
+import { isDeveloper } from '../utils/functions/owner-utils.js';
 
 export const Maintenance: GuardFunction<
   | ArgsOf<'messageCreate' | 'messageReactionAdd' | 'voiceStateUpdate'>
@@ -32,30 +34,31 @@ export const Maintenance: GuardFunction<
   | SimpleCommandMessage
 > = async (argument, _client, next) => {
   const argumentObject = Array.isArray(argument) ? argument[0] : argument;
-  const maintenance = await isInMaintenance();
+  const maintenanceService = container.resolve(MaintenanceService);
+  const maintenance = await maintenanceService.isInMaintenance();
 
   const user =
     argumentObject instanceof CommandInteraction
       ? argumentObject.user
       : argumentObject instanceof MessageReaction
-      ? argumentObject.message.author
-      : argumentObject instanceof VoiceState
-      ? argumentObject.member?.user
-      : argumentObject instanceof Message
-      ? argumentObject.author
-      : argumentObject instanceof SimpleCommandMessage
-      ? argumentObject.message.author
-      : argumentObject instanceof ButtonInteraction ||
-        argumentObject instanceof ChannelSelectMenuInteraction ||
-        argumentObject instanceof CommandInteraction ||
-        argumentObject instanceof ContextMenuCommandInteraction ||
-        argumentObject instanceof MentionableSelectMenuInteraction ||
-        argumentObject instanceof ModalSubmitInteraction ||
-        argumentObject instanceof RoleSelectMenuInteraction ||
-        argumentObject instanceof StringSelectMenuInteraction ||
-        argumentObject instanceof UserSelectMenuInteraction
-      ? argumentObject.member?.user
-      : argumentObject.message?.author;
+        ? argumentObject.message.author
+        : argumentObject instanceof VoiceState
+          ? argumentObject.member?.user
+          : argumentObject instanceof Message
+            ? argumentObject.author
+            : argumentObject instanceof SimpleCommandMessage
+              ? argumentObject.message.author
+              : argumentObject instanceof ButtonInteraction ||
+                  argumentObject instanceof ChannelSelectMenuInteraction ||
+                  argumentObject instanceof CommandInteraction ||
+                  argumentObject instanceof ContextMenuCommandInteraction ||
+                  argumentObject instanceof MentionableSelectMenuInteraction ||
+                  argumentObject instanceof ModalSubmitInteraction ||
+                  argumentObject instanceof RoleSelectMenuInteraction ||
+                  argumentObject instanceof StringSelectMenuInteraction ||
+                  argumentObject instanceof UserSelectMenuInteraction
+                ? argumentObject.member?.user
+                : argumentObject.message?.author;
   if (maintenance && user?.id && !isDeveloper(user.id)) {
     // Make Sure we can reply to the user
     if (
