@@ -1,8 +1,7 @@
-import { Client } from 'discordx';
+import { TextChannel } from 'discord.js';
 
-import { injectable } from 'tsyringe';
+import { inject, injectable } from 'tsyringe';
 
-import { createEncounter } from '../../entities/dt-encounters.mongo.js';
 import {
   gameNPCs,
   GameStatus,
@@ -11,9 +10,11 @@ import {
   RenderPhase,
   renderPhasesArray,
 } from '../../enums/daruma-training.js';
-import type { ChannelSettings } from '../../model/types/daruma-training.js';
-import { DarumaTrainingGameRepository } from '../../repositories/dt-game-repository.js';
-import { getTemporaryPayoutModifier } from '../functions/dt-boost.js';
+import { BoostService } from '../../services/boost-payout.js';
+import { DarumaTrainingEncountersService } from '../../services/dt-encounters.js';
+import { DiscordId } from '../../types/core.js';
+import type { ChannelSettings } from '../../types/daruma-training.js';
+import { generateNPCPlayer } from '../functions/dt-npc-player.js';
 import { phaseDelay } from '../functions/dt-utils.js';
 import logger from '../functions/logger-factory.js';
 
@@ -27,67 +28,64 @@ import { WaitingRoomManager } from './dt-waitingroommanager.js';
  */
 @injectable()
 export class Game {
-  public dtGameRepository: DarumaTrainingGameRepository;
-  public waitingRoomManager: WaitingRoomManager;
-  public embedManager: EmbedManager;
-
-  public state: GameState;
+  private _state?: GameState;
+  private _settings?: ChannelSettings;
   public payoutModifier: number | undefined;
-  public NPC: Player | undefined;
   constructor(
-    public settings: ChannelSettings,
-    public gameRepoInject?: DarumaTrainingGameRepository,
-    public waitingRoomManagerInject?: WaitingRoomManager,
-    public embedManagerInject?: EmbedManager,
-  ) {
-    this.state = new GameState(this);
-
-    this.waitingRoomManager = waitingRoomManagerInject ?? new WaitingRoomManager(this);
-    this.embedManager = embedManagerInject ?? new EmbedManager();
-    this.dtGameRepository = gameRepoInject ?? new DarumaTrainingGameRepository();
+    @inject(WaitingRoomManager) public waitingRoomManager: WaitingRoomManager,
+    @inject(EmbedManager) public embedManager: EmbedManager,
+    @inject(BoostService) public boostService: BoostService,
+    @inject(DarumaTrainingEncountersService)
+    public dtEncountersService: DarumaTrainingEncountersService,
+  ) {}
+  get state(): GameState {
+    if (!this._state) {
+      throw new Error('State has not been initialized yet');
+    }
+    return this._state;
   }
-  async initialize(client: Client): Promise<void> {
-    await this.addNpc();
-    await this.waitingRoomManager.initialize(client);
+
+  get settings(): ChannelSettings {
+    if (!this._settings) {
+      throw new Error('Settings have not been initialized yet');
+    }
+    return this._settings;
   }
   get getNPC(): IGameNPC | undefined {
     return gameNPCs.find((npc) => npc.gameType === this.settings.gameType);
+  }
+  updateState(state: GameState): void {
+    this._state = state;
+  }
+
+  async initialize(channelSettings: ChannelSettings, channel: TextChannel): Promise<void> {
+    this._settings = channelSettings;
+    this.updateState(new GameState(this.settings.token, generateNPCPlayer(this.getNPC)));
+    await this.waitingRoomManager.initialize(this, channel);
   }
   public async addPlayer(player: Player): Promise<boolean> {
     const playerAdded = this.state.playerManager.addPlayer(player);
     await this.embedManager.updateWaitingRoomEmbed(this);
     return playerAdded;
   }
-  public async removePlayer(discordId: string): Promise<boolean> {
+  public async removePlayer(discordId: DiscordId): Promise<boolean> {
     const playerRemoved = this.state.playerManager.removePlayer(discordId);
     await this.embedManager.updateWaitingRoomEmbed(this);
     return playerRemoved;
   }
-  private async addNpc(): Promise<void> {
-    this.NPC = undefined;
-    if (!this.getNPC) {
-      return;
-    }
-    const npcID = this.getNPC.assetIndex;
-    try {
-      this.NPC = await this.dtGameRepository.getNPCPlayer(npcID);
-    } catch (error) {
-      logger.error(`Could not find NPC with ID: ${npcID}`, error);
-    }
-    this.state = this.state.reset();
-  }
   private async endGamePlayerUpdate(): Promise<void> {
-    this.payoutModifier = await getTemporaryPayoutModifier();
-    this.state = this.state.findZenAndWinners(undefined, this.payoutModifier);
+    this.payoutModifier = await this.boostService.getTemporaryPayoutModifier();
+    this.updateState(this.state.findZenAndWinners(undefined, this.payoutModifier));
     for (const player of this.state.playerManager.getAllPlayers()) {
       await player.userAndAssetEndGameUpdate(this.state.gameWinInfo, this.settings.coolDown);
     }
   }
   private async saveEncounter(): Promise<number> {
     await this.endGamePlayerUpdate();
-    return await createEncounter(this);
+    const { gameType, channelId } = this.settings;
+    const allPlayers = this.state.playerManager.getAllPlayers();
+    return await this.dtEncountersService.create(allPlayers, channelId, gameType);
   }
-
   async startChannelGame(): Promise<void> {
     await this.startGame();
     await this.handleGameLogic();
@@ -95,7 +93,7 @@ export class Game {
   }
   private async startGame(): Promise<void> {
     const encounterId = await this.saveEncounter();
-    this.state = this.state.startGame(encounterId);
+    this.updateState(this.state.startGame(encounterId));
     await this.embedManager.startGame(this);
   }
   private async handleGameLogic(
@@ -109,7 +107,7 @@ export class Game {
       const allPlayers = this.state.playerManager.getAllPlayers();
       while (this.state.status === GameStatus.activeGame) {
         for (const [index, player] of allPlayers.entries()) {
-          this.state = this.state.setCurrentPlayer(player, index);
+          this.updateState(this.state.setCurrentPlayer(player, index));
 
           for (const phase of renderPhasesArray) {
             const board = this.state.renderThisBoard(phase);
@@ -118,11 +116,11 @@ export class Game {
           }
         }
 
-        this.state = this.state.nextRoll();
+        this.updateState(this.state.nextRoll());
       }
     } catch (error) {
       logger.error(`Error in gameHandler: ${JSON.stringify(error)}`);
-      this.state = this.state.updateStatus(GameStatus.win);
+      this.updateState(this.state.updateStatus(GameStatus.win));
     }
   }
   private async finishGame(): Promise<void> {
