@@ -1,32 +1,24 @@
-import { Client, Message, MessageCreateOptions, MessagePayload, TextChannel } from 'discord.js';
+import { Message, MessageCreateOptions, MessagePayload, TextChannel } from 'discord.js';
+
+import { injectable } from 'tsyringe';
 
 import { GameStatus } from '../../enums/daruma-training.js';
-import { removeChannelFromDatabase } from '../../repositories/dt-channel-repository.js';
 import logger from '../functions/logger-factory.js';
 
 import { Game } from './dt-game.js';
 
+@injectable()
 export class WaitingRoomManager {
-  private game: Game;
   public waitingRoomChannel?: TextChannel;
-
-  constructor(game: Game) {
+  public game?: Game;
+  constructor() {}
+  async initialize(game: Game, channel: TextChannel): Promise<void> {
     this.game = game;
-  }
-  async initialize(client: Client): Promise<void> {
-    try {
-      this.waitingRoomChannel = (await client.channels.fetch(
-        this.game.settings.channelId,
-      )) as TextChannel;
-    } catch {
-      logger.error(`Could not find channel ${this.game.settings.channelId} -- Removing from DB`);
-      await removeChannelFromDatabase(this.game.settings.channelId);
-      return;
-    }
+    this.waitingRoomChannel = channel;
+    await this.game.embedManager.sendJoinWaitingRoomEmbed(this.game);
     logger.info(
       `Channel ${this.waitingRoomChannel.name} (${this.waitingRoomChannel.id}) of type ${this.game.settings.gameType} has been started`,
     );
-    await this.game.embedManager.sendJoinWaitingRoomEmbed(this.game);
   }
   async sendToChannel(
     content: string | MessagePayload | MessageCreateOptions,
@@ -35,7 +27,7 @@ export class WaitingRoomManager {
   }
 
   async stopWaitingRoomOnceGameEnds(): Promise<void> {
-    if (this.game.state.status === GameStatus.waitingRoom) {
+    if (this.game?.state.status === GameStatus.waitingRoom) {
       await this.game.embedManager.sendJoinWaitingRoomEmbed(this.game);
     } else {
       return;
