@@ -1,59 +1,48 @@
 import { createLogger, format, Logger, transports } from 'winston';
-import type * as Transport from 'winston-transport';
 
-const { combine, splat, timestamp, colorize, printf } = format;
-const isTestEnvironment = process.env?.['NODE_ENV'] === 'test';
-
-class JestFilterTransport extends transports.Console {
-  public override log(info: unknown, callback: () => void): void {
-    /* istanbul ignore next */
-    if (!isTestEnvironment && super.log) {
-      super.log(info, callback);
-    } else {
-      setImmediate(callback);
-    }
-  }
-}
-
-export function createLoggerFactory(level: string): Logger {
-  const consoleTransport = createConsoleTransport(level);
+export function createLoggerFactory(
+  loggerName: string = '',
+  handleExceptions: boolean = false,
+  node_environment?: string,
+): Logger {
+  const loggerEnvironment = node_environment || process.env['NODE_ENV'];
+  const isTestEnvironment = loggerEnvironment === 'test';
+  const production = loggerEnvironment === 'production';
   /* istanbul ignore next */
-  const transportsArray = isTestEnvironment ? [createJestFilterTransport()] : [consoleTransport];
-
-  return createLogger({
-    level,
-    transports: transportsArray,
-    exitOnError: false,
-  });
-}
-
-function createConsoleTransport(level: string): Transport {
-  return new transports.Console({
-    level,
-    format: combine(colorize(), splat(), timestamp(), createLogFormat()),
-  });
-}
-
-function createJestFilterTransport(): Transport {
-  return new JestFilterTransport({
-    level: 'debug',
-    format: combine(splat(), timestamp(), createLogFormat()),
-    handleExceptions: true,
-    handleRejections: true,
-  });
-}
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function createLogFormat(): any {
-  return printf(({ level, message, timestamp, ...metadata }) => {
-    // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-    let message_ = `⚡ ${timestamp} [${level}] : ${message} `;
+  let logFormat = format.printf(({ level, message, timestamp, logger, ...metadata }) => {
+    let message_ = `⚡ ${timestamp} ${logger ? `[${logger}] ` : ''}[${level}] : ${message} `;
     if (metadata && Object.keys(metadata).length > 0) {
       message_ += JSON.stringify(metadata);
     }
     return message_;
   });
+  let logLevel = 'debug';
+  if (loggerName === 'DiscordX') {
+    logFormat = format.printf(({ message }) => {
+      return message;
+    });
+  }
+  if (loggerName === 'GlobalEmitter') {
+    logLevel = production ? 'none' : 'debug';
+  }
+  const transport = new transports.Console({
+    level: logLevel,
+    handleExceptions,
+    format: format.combine(format.colorize(), format.splat(), format.timestamp(), logFormat),
+  });
+
+  return createLogger({
+    level: logLevel,
+    silent: isTestEnvironment,
+    transports: [transport],
+    exitOnError: false,
+    defaultMeta: { logger: loggerName },
+  });
 }
 
-const logger = createLoggerFactory('debug');
+// Exported loggers
+export const globalEmitterLogger = createLoggerFactory('GlobalEmitter');
 
-export default logger;
+export const discordXLogger = createLoggerFactory('DiscordX');
+
+export default createLoggerFactory('', true);
